owners can edit or delete

+++ src/events/events.controller.ts
@@ -14,12 +14,13 @@ import {
   UsePipes,
   ValidationPipe,
   UseGuards,
+  ForbiddenException,
 } from '@nestjs/common'
 import { CreateEventDto } from './input/create-event.dto'
 import { UpdateEventDto } from './input/update-event.dto'
 import { Event } from './event.entity'
 import { InjectRepository } from '@nestjs/typeorm'
-import { Like, MoreThan, Repository } from 'typeorm'
+import { Repository } from 'typeorm'
 import { Attendee } from './attendee.entity'
 import { EventService } from './events.service'
 import { ListEvents } from './list.events'
@@ -54,17 +55,17 @@ export class EventsController {
 
   @Get('/practice')
   async practice() {
-    return await this.repository.find({
-      select: ['id', 'when'],
-      where: [
-        { id: MoreThan(3), when: MoreThan(new Date('2021-02-12T13:00:00')) },
-        { description: Like('%meet%') },
-      ],
-      take: 2,
-      order: {
-        id: 'DESC',
-      },
-    })
+    // return await this.repository.find({
+    //   select: ['id', 'when'],
+    //   where: [
+    //     { id: MoreThan(3), when: MoreThan(new Date('2021-02-12T13:00:00')) },
+    //     { description: Like('%meet%') },
+    //   ],
+    //   take: 2,
+    //   order: {
+    //     id: 'DESC',
+    //   },
+    // })
   }
 
   @Get('/practice2')
@@ -102,23 +103,43 @@ export class EventsController {
   }
 
   @Patch(':id')
-  async update(@Param('id') id, @Body() input: UpdateEventDto) {
-    const event = await this.repository.findOneBy({ id })
+  @UseGuards(AuthGuardJwt)
+  async update(
+    @Param('id') id,
+    @Body() input: UpdateEventDto,
+    @CurrentUser() user: User,
+  ) {
+    const event = await this.eventService.getEvent(id)
 
-    return await this.repository.save({
-      ...event,
-      ...input,
-      when: input.when ? new Date(input.when) : event.when,
-    })
+    if (!event) {
+      throw new NotFoundException()
+    }
+
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `You are not authorized to change this event`,
+      )
+    }
+    return await this.eventService.updateEvent(event, input)
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(204)
-  async remove(@Param('id') id) {
-    const result = await this.eventService.deleteEvent(id)
-
-    if (result?.affected === 0) {
+  async remove(@Param('id') id, @CurrentUser() user: User) {
+    const event = await this.eventService.getEvent(id)
+    if (!event) {
       throw new NotFoundException()
     }
+
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `You are not authorized to delete this event`,
+      )
+    }
+
+    return await this.eventService.deleteEvent(id)
   }
 }
